Refactor some test utility functions into their own file
